Refactor IOJ Appeal creation to return API response directly
Updated `IojAppealDualWriteService` to return `ApiCreateIojAppealResponse` instead of `IojAppealEntity`. Simplified `IojAppealController` by removing redundant response building logic. Adjusted tests to align with the refactored service method.
