+++ crime-assessment-service/src/main/java/uk/gov/justice/laa/crime/assessmentservice/iojappeal/controller/IojAppealController.java
@@ -3,8 +3,7 @@ package uk.gov.justice.laa.crime.assessmentservice.iojappeal.controller;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import uk.gov.justice.laa.crime.assessmentservice.common.exception.CrimeValidationException;
-import uk.gov.justice.laa.crime.assessmentservice.iojappeal.entity.IojAppealEntity;
+import uk.gov.justice.laa.crime.assessmentservice.common.api.exception.CrimeValidationException;
 import uk.gov.justice.laa.crime.assessmentservice.iojappeal.service.IojAppealDualWriteService;
 import uk.gov.justice.laa.crime.assessmentservice.iojappeal.validator.ApiCreateIojAppealRequestValidator;
 import uk.gov.justice.laa.crime.common.model.ioj.ApiCreateIojAppealRequest;
@@ -56,11 +55,8 @@ public class IojAppealController implements IojAppealApi {
             throw new CrimeValidationException(validationErrors);
         }
 
-        IojAppealEntity appealEntity = iojAppealDualWriteService.createIojAppeal(request);
+        ApiCreateIojAppealResponse iojAppeal = iojAppealDualWriteService.createIojAppeal(request);
 
-        ApiCreateIojAppealResponse response = new ApiCreateIojAppealResponse()
-                .withAppealId(appealEntity.getAppealId().toString())
-                .withLegacyAppealId(appealEntity.getLegacyAppealId());
-        return ResponseEntity.ok(response);
+        return ResponseEntity.ok(iojAppeal);
     }
 }
