+++ crime-assessment-service/src/main/java/uk/gov/justice/laa/crime/assessmentservice/iojappeal/service/IojAppealDualWriteService.java
@@ -3,7 +3,7 @@ package uk.gov.justice.laa.crime.assessmentservice.iojappeal.service;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import uk.gov.justice.laa.crime.assessmentservice.audit.api.IojAudit;
-import uk.gov.justice.laa.crime.assessmentservice.common.exception.AssessmentRollbackException;
+import uk.gov.justice.laa.crime.assessmentservice.common.api.exception.AssessmentRollbackException;
 import uk.gov.justice.laa.crime.assessmentservice.iojappeal.entity.IojAppealEntity;
 import uk.gov.justice.laa.crime.common.model.ioj.ApiCreateIojAppealRequest;
 import uk.gov.justice.laa.crime.common.model.ioj.ApiCreateIojAppealResponse;
@@ -49,7 +49,7 @@ public class IojAppealDualWriteService {
     }
 
     @Transactional
-    public IojAppealEntity createIojAppeal(ApiCreateIojAppealRequest request) {
+    public ApiCreateIojAppealResponse createIojAppeal(ApiCreateIojAppealRequest request) {
         IojAppealEntity appealEntity = iojAppealService.create(request);
         ApiCreateIojAppealResponse legacyAppeal = legacyIojAppealService.create(request);
         Integer legacyAppealId = legacyAppeal.getLegacyAppealId();
@@ -67,6 +67,8 @@ public class IojAppealDualWriteService {
                     appealEntity.getAppealId().toString(), legacyAppealId, e.getMessage()));
         }
 
-        return appealEntity;
+        return new ApiCreateIojAppealResponse()
+                .withAppealId(appealEntity.getAppealId().toString())
+                .withLegacyAppealId(legacyAppealId);
     }
 }
